Move `block_deserializer` to `transport` namespace
